requirements.txt
langchain==1.2.2
langchain-community==0.4.1
langchain-core==1.2.6
langchain-huggingface==1.2.0
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langchain-classic==1.0.1
chromadb==1.4.0
faiss-cpu==1.10.0
pypdf==5.1.0
gradio==6.2.0
sentence-transformers==5.2.0
python-dotenv==1.2.1
